pyproject.toml
[project]
name = "pa-translator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "groq>=0.28.0",
    "python-dotenv>=1.1.1",
]
